requirements.txt
yfinance>=0.2.0
pandas>=1.3.0
numpy>=1.21.0
ta>=0.10.0
mplfinance>=0.12.0
requests>=2.26.0
tweepy>=4.14.0
apscheduler>=3.10.0
pytz>=2023.0
pyyaml>=6.0
schedule>=1.1.0
